+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Require JobRouter Client in version 2
 - Require JobRouter Connector extension in version 2
 - Require JobRouter Base extension in version 2
+- Constants from Table model are moved into the new TableType enum
 
 ### Fixed
 - Error in content element when table link is disabled or deleted