+++ Documentation/changelog.rst
@@ -18,6 +18,7 @@ Changed
 * Require JobRouter Client in version 2
 * Require JobRouter Connector extension in version 2
 * Require JobRouter Base extension in version 2
+* Constants from Table model are moved into the new TableType enum
 
 Fixed
 ^^^^^